src/requirements.txt
numpy==1.21.2
pandas==1.3.3
tbats==1.1.0
pmdarima==1.8.2
snowflake-connector-python==2.7.3
python-dotenv==0.19.0
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
mlflow==1.20.2